requirements.txt
Pillow==8.2.0
pytesseract==0.3.8
requests==2.26.0
beautifulsoup4==4.10.0
